Re-organize schema crate migration defs
and start scaffolding manual migration plan defs
