+++ crates/client-api/src/routes/database.rs
@@ -45,7 +45,7 @@ use spacetimedb_lib::db::raw_def::v10::RawModuleDefV10;
 use spacetimedb_lib::db::raw_def::v9::RawModuleDefV9;
 use spacetimedb_lib::http as st_http;
 use spacetimedb_lib::{sats, AlgebraicValue, Hash, ProductValue, Timestamp};
-use spacetimedb_schema::auto_migrate::{
+use spacetimedb_schema::migrate::{
     MigrationPolicy as SchemaMigrationPolicy, MigrationToken, PrettyPrintStyle as AutoMigratePrettyPrintStyle,
 };
 use tokio::sync::oneshot;
