+++ crates/schema/src/lib.rs
@@ -2,10 +2,10 @@
 //!
 //! Handles validation and normalization of raw schema definitions from the `spacetimedb_lib` crate.
 
-pub mod auto_migrate;
 pub mod def;
 pub mod error;
 pub mod identifier;
+pub mod migrate;
 pub mod reducer_name;
 pub mod relation;
 pub mod schema;
